services/user_management/requirements.txt
kink~=0.6.5
loguru~=0.6.0
fastapi~=0.78.0
uvicorn~=0.17.6
uvicorn[standard]
pytz~=2022.1
pydantic~=1.9.1
confluent-kafka
pytest~=7.1.2
pytest-asyncio
py-bcrypt
sqlalchemy~=1.4.40

bcrypt~=4.0.0
python-dotenv~=0.20.0
asyncpg